Supporting ocidir GC in regclient
